use helpers to avoid re-runnning checks
